Reset HTML and JS code to begin rewrite of project
Previous approach to the project was inefficient, specifically relating
to the use of conditionals and the merging of JS-based operations with
the HTML/CSS-based operations.

This approach also caused issues when new features were added and so
the best solution was to restart the project and build a solution that
avoids such problems arising.